Divided the sections into separate js components, deleted unused variables from each js, need to clean css
